Move cycle detection to `utils` package.
