+++ packages/utils/src/cycle-detection.ts
